requirements.txt
streamlit
faiss-cpu
sentence-transformers
python-dotenv
PyMuPDF
requests
